Revert "Major refactor of grabABI in prep for fixing bugs"
This reverts commit d8960efecc9bc584a0421cb3ae1ce20cfd0af49c.
