+++ src/tools/grabABI/options.h
@@ -17,14 +17,16 @@ class COptions : public COptionsBase {
 public:
     uint64_t parts;
     bool noconst;
+    bool open;
+    bool asJson;
     bool raw;
     bool asData;
     bool decNames;
-    bool isGenerate;
     CAddressArray addrs;
     string_q primaryAddr;
     string_q classDir;
     string_q prefix;
+    string_q theABI;
     bool silent;
 
     COptions(void);
@@ -39,8 +41,4 @@ public:
     bool isBuiltin(void) const { return isToken() || isWallet(); }
 };
 
-//-----------------------------------------------------------------------
-extern string_q getPrefix       (const string_q& in);
-extern string_q acquireABI      (CFunctionArray& functions, const address_t& addr, const COptions& opt, bool builtIn);
-extern void     handle_display  (COptions& options);
-extern void     handle_generate (COptions& options);
+extern string_q getPrefix(const string_q& in);
