+++ src/tools/grabABI/templates/CMakeLists.txt
@@ -28,9 +28,6 @@ set(TEST_EXE "${EXECUTABLE_OUTPUT_PATH}/${TOOL_NAME}")
 set(TEST_PATH "${TEST_PATH}/${PROJ_NAME}/${TOOL_NAME}")
 set(GOLD_PATH "${GOLD_PATH}/${PROJ_NAME}/${TOOL_NAME}")
 
-# To run only specific tests we define a test configuration named "test_all"
-add_custom_target(test-all COMMAND ${CMAKE_CTEST_COMMAND} -C test_all)
-
 # Function to run an special or slow test case
 function(run_special_test testName)
     run_the_special_test(${TEST_PATH} ${GOLD_PATH} ${testName} ${TEST_EXE} ${ARGN})
