+++ test/gold/tools/grabABI/grabABI_invalid_1.txt
@@ -1,7 +1,7 @@
 grabABI argc: 2 [1:0x00001] 
 grabABI 0x00001 
 
-  Invalid address '0x00001'. Length (7) is not equal to 40 characters (20 bytes).
+  Invalid address '0x00001'. Length is not equal to 40 characters (20 bytes).
 
 
   Usage:    grabABI [-c|-g|-d|-e|-n|-o|-r|-v|-h] addr
