--- test/gold/tools/grabABI/grabABI_known_contracts.txt
@@ -1,52 +1,0 @@
-grabABI argc: 4 [1:0xTokenLib] [2:0xWalletLib] [3:-e] 
-grabABI 0xTokenLib 0xWalletLib -e 
-ABI for address 0xTokenLib and others
-	function name                       () [0x06fdde03]
-	function approve                    (address _spender,uint256 _value) [0x095ea7b3]
-	function totalSupply                () [0x18160ddd]
-	function transferFrom               (address _from,address _to,uint256 _value) [0x23b872dd]
-	function decimals                   () [0x313ce567]
-	function version                    () [0x54fd4d50]
-	function balanceOf                  (address _owner) [0x70a08231]
-	function symbol                     () [0x95d89b41]
-	function transfer                   (address _to,uint256 _value) [0xa9059cbb]
-	function approveAndCall             (address _spender,uint256 _value,bytes _extraData) [0xcae9ca51]
-	function allowance                  (address _owner,address _spender) [0xdd62ed3e]
-	function ownerOf                    (uint256 _tokenId) [0x6352211e]
-	function exists                     (uint256 _tokenId) [0x4f558e79]
-	function getApproved                (uint256 _tokenId) [0x081812fc]
-	function setApprovalForAll          (address _operator,bool _approved) [0xa22cb465]
-	function isApprovedForAll           (address _owner,address _operator) [0xe985e9c5]
-	function safeTransferFrom           (address _from,address _to,uint256 _tokenId) [0x42842e0e]
-	function safeTransferFrom           (address _from,address _to,uint256 _tokenId,bytes _data) [0xb88d4fde]
-	function supportsInterface          (bytes4 _interfaceID) [0x01ffc9a7]
-	function onERC721Received           (address _operator,address _from,uint256 _tokenId,bytes _data) [0x150b7a02]
-	event Transfer                      (address indexed _from,address indexed _to,uint256 _value) [0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef]
-	event Approval                      (address indexed _owner,address indexed _spender,uint256 _value) [0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925]
-	event ApprovalForAll                (address indexed _owner,address indexed _operator,bool _approved) [0x17307eab39ab6107e8899845ad3d59bd9653f200f220920489ca2b5937696c31]
-	function removeOwner                (address _owner) [0x173825d9]
-	function isOwner                    (address _addr) [0x2f54bf6e]
-	function m_numOwners                () [0x4123cb6b]
-	function resetSpentToday            () [0x5c52c2f5]
-	function addOwner                   (address _owner) [0x7065cb48]
-	function m_required                 () [0x746c9171]
-	function confirm                    (bytes32 _h) [0x797af627]
-	function setDailyLimit              (uint256 _newLimit) [0xb20d30a9]
-	function execute                    (address _to,uint256 _value,bytes _data) [0xb61d27f6]
-	function revoke                     (bytes32 _operation) [0xb75c7dc6]
-	function changeRequirement          (uint256 _newRequired) [0xba51a6df]
-	function hasConfirmed               (bytes32 _operation,address _owner) [0xc2cf7326]
-	function kill                       (address _to) [0xcbf0b0c0]
-	function changeOwner                (address _from,address _to) [0xf00d4b5d]
-	function m_dailyLimit               () [0xf1736d86]
-	event Confirmation                  (address owner,bytes32 operation) [0xe1c52dc63b719ade82e8bea94cc41a0d5d28e4aaf536adb5e9cccc9ff8c1aeda]
-	event Revoke                        (address owner,bytes32 operation) [0xc7fb647e59b18047309aa15aad418e5d7ca96d173ad704f1031a2c3d7591734b]
-	event OwnerChanged                  (address oldOwner,address newOwner) [0xb532073b38c83145e3e5135377a08bf9aab55bc0fd7c1179cd4fb995d2a5159c]
-	event OwnerAdded                    (address newOwner) [0x994a936646fe87ffe4f1e469d3d6aa417d6b855598397f323de5b449f765f0c3]
-	event OwnerRemoved                  (address oldOwner) [0x58619076adf5bb0943d100ef88d52d7c3fd691b19d3a9071b555b651fbf418da]
-	event RequirementChanged            (uint256 newRequirement) [0xacbdb084c721332ac59f9b8e392196c9eb0e4932862da8eb9beaf0dad4f550da]
-	event Deposit                       (address from,uint256 value) [0xe1fffcc4923d04b559f4d29a8bfc6cda04eb5b0d3c460751c2402c5c5cc9109c]
-	event SingleTransact                (address owner,uint256 value,address to,bytes data) [0x92ca3a80853e6663fa31fa10b99225f18d4902939b4c53a9caae9043f6efd004]
-	event MultiTransact                 (address owner,bytes32 operation,uint256 value,address to,bytes data) [0xe7c957c06e9a662c1a6c77366179f5b702b97651dc28eee7d5bf1dff6e40bb4a]
-	event ConfirmationNeeded            (bytes32 operation,address initiator,uint256 value,address to,bytes data) [0x1733cbb53659d713b79580f79f3f9ff215f78a7c7aa45890f3b89fc5cddfbf32]
-
